[Mosaic GPU] Unify `evaluate_equation` and `simplify_equation` into `reduce_equation`.
In an upcoming change, we will need to allow `simplify_expression` to return
`Unsatisfiable`. If simplification can now make an equation system
unsatisfiable, it makes more sense to call it `reduce` or `evaluate` as well
now.

As a result, we also can no longer call `reduce_equation` in an
`EquationSystem`'s `__post_init__`.

PiperOrigin-RevId: 778503961